// DATN/penguin-hospital/src/app/model/class/medicalReport.ts
export class MedicalReport{
  code: string ='';
  namePet: string = '';
  oldPet: string = '';
  weightPet: string = '';
  genderPet: string = 'default';
  symptom: string ='';
  numberPhone: string = '';
  address: string = '';
  petOwner: string = '';
  followSchedule: string = '';
  diagnosed : string ='';
  appointmentId: number = 0;
  urlPet:string = '';
}